+++ src/clj/rems/api/services/form.clj
@@ -2,6 +2,7 @@
   (:require [clojure.test :refer :all]
             [com.rpl.specter :refer [ALL transform]]
             [medley.core :refer [filter-vals]]
+            [rems.api.services.dependencies :as dependencies]
             [rems.api.services.util :as util]
             [rems.api.services.workflow :as workflow]
             [rems.api.schema :refer [FieldTemplate]]
@@ -15,28 +16,8 @@
             [schema.core :as s])
   (:import rems.InvalidRequestException))
 
-(defn- catalogue-items-for-form [id include-archived]
-  (->> (catalogue/get-localized-catalogue-items {:form id :archived include-archived})
-       (map #(select-keys % [:id :title :localizations]))))
-
-(defn- workflows-for-form [id include-archived]
-  ;; TODO optimize?
-  (->> (workflow/get-workflows (when-not include-archived
-                                 {:archived false}))
-       (filter #(contains? (set (get-in % [:workflow :forms])) {:form/id id}))
-       (map #(select-keys % [:id :title]))))
-
-(defn- form-in-use-error [form-id include-archived]
-  (let [catalogue-items (seq (catalogue-items-for-form form-id include-archived))
-        workflows (seq (workflows-for-form form-id include-archived))]
-    (when (or catalogue-items workflows)
-      {:success false
-       :errors [{:type :t.administration.errors/form-in-use
-                 :catalogue-items catalogue-items
-                 :workflows workflows}]})))
-
 (defn form-editable [form-id]
-  (or (form-in-use-error form-id true)
+  (or (dependencies/in-use-error {:form/id form-id})
       {:success true}))
 
 (defn validate-given-ids
@@ -109,6 +90,8 @@
                                                     :title (:form/title form)
                                                     :user user-id
                                                     :fields (serialize-fields form)}))]
+          ;; reset-cache! not strictly necessary since forms don't depend on anything, but here for consistency
+          (dependencies/reset-cache!)
           {:success (not (nil? form-id))
            :id form-id}))))
 
@@ -131,7 +114,7 @@
     ;; need to check both previous and new organization
     (util/check-allowed-organization! (:form/organization (get-form-template form-id)))
     (util/check-allowed-organization! organization)
-    (or (form-in-use-error form-id true)
+    (or (dependencies/in-use-error {:form/id form-id})
         (validation-error form)
         (do (db/edit-form-template! {:id form-id
                                      :organization (:organization/id organization)
@@ -147,9 +130,8 @@
 
 (defn set-form-archived! [{:keys [id archived]}]
   (util/check-allowed-organization! (:form/organization (get-form-template id)))
-  (let [catalogue-items (catalogue-items-for-form id false)]
-    (or (and archived (form-in-use-error id false))
-        (do
-          (db/set-form-template-archived! {:id id
-                                           :archived archived})
-          {:success true}))))
+  (or (dependencies/change-archive-status-error archived {:form/id id})
+      (do
+        (db/set-form-template-archived! {:id id
+                                         :archived archived})
+        {:success true})))