+++ src/clj/rems/api/services/licenses.clj
@@ -1,6 +1,7 @@
 (ns rems.api.services.licenses
   "Serving licenses for API."
-  (:require [rems.api.services.util :as util]
+  (:require [rems.api.services.dependencies :as dependencies]
+            [rems.api.services.util :as util]
             [rems.common.util :refer [distinct-by]]
             [rems.db.applications :as applications]
             [rems.db.attachments :as attachments]
@@ -24,6 +25,8 @@
                                         :title (:title localization)
                                         :textcontent (:textcontent localization)
                                         :attachmentId (:attachment-id localization)}))
+    ;; reset-cache! not strictly necessary since licenses don't depend on anything, but here for consistency
+    (dependencies/reset-cache!)
     {:success (not (nil? licid))
      :id licid}))
 
@@ -58,22 +61,6 @@
         (when-let [attachment (get-license-attachment attachment-id)]
           attachment)))))
 
-(defn- get-license-usage [id]
-  ;; these could be db joins
-  (let [resources (->> (db/get-resources-for-license {:id id})
-                       (map :resid)
-                       (map resource/get-resource)
-                       (remove :archived)
-                       (map #(select-keys % [:id :resid])))
-        workflows (->> (db/get-workflows-for-license {:id id})
-                       (map :wfid)
-                       (map workflow/get-workflow)
-                       (remove :archived)
-                       (map #(select-keys % [:id :title])))]
-    (when (or (seq resources) (seq workflows))
-      {:resources resources
-       :workflows workflows})))
-
 (defn get-license
   "Get a single license by id"
   [id]
@@ -87,15 +74,11 @@
 
 (defn set-license-archived! [{:keys [id archived]}]
   (util/check-allowed-organization! (:organization (get-license id)))
-  (let [usage (get-license-usage id)]
-    (if (and archived usage)
-      {:success false
-       :errors [(merge {:type :t.administration.errors/license-in-use}
-                       usage)]}
+  (or (dependencies/change-archive-status-error archived  {:license/id id})
       (do
         (db/set-license-archived! {:id id
                                    :archived archived})
-        {:success true}))))
+        {:success true})))
 
 (defn get-all-licenses
   "Get all licenses.