+++ src/cljs/rems/administration/workflow.cljs
@@ -62,7 +62,12 @@
                                                                          (map enrich-user)
                                                                          (map :display)
                                                                          (str/join ", "))]
-              [inline-info-field (text :t.administration/active) [readonly-checkbox {:value (status-flags/active? workflow)}]]]}]
+              [inline-info-field (text :t.administration/active) [readonly-checkbox {:value (status-flags/active? workflow)}]]
+              [inline-info-field (text :t.administration/forms)
+               (into [:ul.list-group]
+                     (for [form (get-in workflow [:workflow :forms])]
+                       [:li.list-group-item [atoms/link nil (str "/administration/forms/" (:form/id form)) (:form/title form)]]))
+               {:box? false}]]}]
    [licenses-view (:licenses workflow) language]
    (let [id (:id workflow)]
      [:div.col.commands