+++ test/clj/rems/api/test_forms.clj
@@ -222,9 +222,8 @@
                                                       :organization {:organization/id "organization1"}})]
         (testing "Form is non-editable after in use by a catalogue item"
           (is (= {:success false
-                  :errors [{:type "t.administration.errors/form-in-use"
-                            :catalogue-items [{:id cat-id :localizations {}}]
-                            :workflows nil}]}
+                  :errors [{:type "t.administration.errors/in-use-by"
+                            :catalogue-items [{:id cat-id :localizations {}}]}]}
                  (api-call :get (str "/api/forms/" form-id "/editable") nil
                            api-key user-id)))
           (testing "even if catalogue item is archived & disabled"
@@ -233,11 +232,10 @@
             (api-call :put "/api/catalogue-items/enabled" {:id cat-id :enabled false}
                       api-key user-id)
             (is (= {:success false
-                  :errors [{:type "t.administration.errors/form-in-use"
-                            :catalogue-items [{:id cat-id :localizations {}}]
-                            :workflows nil}]}
-                 (api-call :get (str "/api/forms/" form-id "/editable") nil
-                           api-key user-id)))))))
+                    :errors [{:type "t.administration.errors/in-use-by"
+                              :catalogue-items [{:id cat-id :localizations {}}]}]}
+                   (api-call :get (str "/api/forms/" form-id "/editable") nil
+                             api-key user-id)))))))
     (let [form-id (:id (api-call :post "/api/forms/create"
                                  {:form/organization {:organization/id "organization1"}
                                   :form/title "form editable test 2"
@@ -250,8 +248,7 @@
         (let [wfid (test-data/create-workflow! {:forms [{:form/id form-id}]
                                                 :title "wf with form"})]
           (is (= {:success false
-                  :errors [{:type "t.administration.errors/form-in-use"
-                            :catalogue-items nil
+                  :errors [{:type "t.administration.errors/in-use-by"
                             :workflows [{:id wfid :title "wf with form"}]}]}
                  (api-call :get (str "/api/forms/" form-id "/editable") nil
                            api-key user-id)))
@@ -261,8 +258,7 @@
             (api-call :put "/api/workflows/enabled" {:id wfid :enabled false}
                       api-key user-id)
             (is (= {:success false
-                    :errors [{:type "t.administration.errors/form-in-use"
-                              :catalogue-items nil
+                    :errors [{:type "t.administration.errors/in-use-by"
                               :workflows [{:id wfid :title "wf with form"}]}]}
                    (api-call :get (str "/api/forms/" form-id "/editable") nil
                              api-key user-id)))))))))